Tries to merge data retrieval with node generation
This contains a few defined functions for generating the initial
graph and a mock function for determining the zoom. The former
has been improved to allow a user to specify the region via the command
line. The zoom feature is not fully implemented. A placeholder funciton that
includes a partial implementation is present here.

Once the graph window is closed, an image is saved to the images directory. This
image is always called Temp.png, so we can reliably hard code this path without
breaking the program.

I've also added a .gitignore to prevent auto-generated folders and files from
being included in the repository.

NOTE: I could not perform testing because the node generation crashed my
computer.