Optimization.py
This file attemtps to maximize performance, while at the same time
reducing the memory footprint of the original program
(TestScenario3.py).

The pygame library has been stripped out in favor of using only
Matplotlib. Matplotlib.collections are used to speed up the drawing
process. No optimization was done to the calculation of origins, radii,
or iteration through the image.

I found an [article] (https://pythonprogramming.net/how-to-embed-matplotlib-graph-tkinter-gui/)
on integrating plots into tkiniter. This would lea dto zero delay between
the user seeing the graph and seeing the cost of the nodes. Solving this issue would
greatly improve the usability of our applicaiton.