I fixed the heatmap!
The heatmap now works. The main issue was that I added the
patches to the axes before running them through PatchCollection(),
which sets specific properties. Also, the match_original paramter
for PatchCollection() defaulted to false.

Due to matplotlib's optimizations when drawing
PatchCollections, we can specify thousands of circles and it'll still
take only a few seconds to draw all of them. But, having so many circles
draw on top of each other results in the streets being covered. For the
time being, I've commented out all but two circles that are generated
per node. We could either add a couple more or better segment the
heatmap so each node doesn't contain as many subdivisions.