Skips makeGraph() call if the graph is present
makeGraph() take a considerable amount of time to run because it has to
download the OSM data from the internet. Thus, we should only run
this function if it's absolutely necessary. During makeGraph() the
graph's projection is now saved in a .gml file. We only run makeGraph()
if the .gml file does not exist. If it does, we assume that we can read
from it and calculate the square meters as we did inside makeGraph(),
with considerably less wait time.

Next, we use a try-catch when attempting to read the image using PIL.
It's entirely possible that a user might have deleted the city's .png
file from the images directory without removing the corresponding
.graphml file from the data directory. In this case, we have no choice
but to run makeGraph() in order to create both the .gml and .png files.